Move testing utilities to a top-level directory
The current layout has a lot of annoying deep nesting. Ideally we'd move lib/shared to the top level, but ng-packagr doesn't seem to allow that; everything that's a runtime dependency of the library has to be under lib, since that's where lib's package.json is. But this restriction does not apply to testing utilities, since we don't ship those as part of the library, so we can reduce the amount of nesting by moving those to the top level.
